Add instructions: LDXA, STXD, LDX
Note that STXD is postdecrement and LDXA is postincrement. Therefore
STXD and LDXA cannot naively be thought of as pair of symmetric PUSH and
POP instructions. The Programming Techniques section of the CDP1802 User
Manual which shows the following example code for pushing and popping
R3, confirming non-symmetric interpretation of the semantics:

    GHI  R3
    STXD
    GLO  R3
    STXD

    IRX
    LDXA
    PLO  R3
    LDX
    PHI  R3